critcom_agent/agent.py
"""
CritCom — Critical Results Communication Agent (FHIR-only build).

Deployed on Prompt Opinion's Agent Assemble platform.
Uses Google ADK + A2A. FHIR context is injected automatically by
Prompt Opinion via the before_model_callback hook.

DICOM worklist path is intentionally excluded from this build.
"""

from __future__ import annotations

import os

from shared.fhir_hook import make_extract_fhir_context
from shared.tools import ALL_TOOLS

INSTRUCTION = """\
You are CritCom, a critical results communication agent for radiology.

Your job: when a radiologist signs a report containing a critical finding, ensure
the right physician is notified, the notification is tracked in FHIR, and if no
acknowledgment is received within the required timeframe, escalate to the on-call
backup provider.

You receive patient context (FHIR URL, FHIR token, patient ID) automatically from
the Prompt Opinion platform. All FHIR calls use these injected credentials.

When you are asked to process a study or report:

1. Fetch the report using fetch_report_fhir_tool. Pass either:
   - diagnostic_report_id  (preferred — the DiagnosticReport resource ID)
   - service_request_id    (fallback — the linked ServiceRequest ID)
   If neither is provided and you have a patient_id from session state, search
   by patient to find the most recent final DiagnosticReport.

2. Read the returned study's acr_category:
   - If the DiagnosticReport had an ACR tag set by the RIS, it will be present.
   - If not, the tool automatically runs LLM classification on the report text.
   - If acr_category is "Cat3", "None", or null after fetching, stop and report
     that no critical communication is needed.

3. For Cat1 or Cat2: call resolve_provider_tool with the service_request_id to
   find the ordering physician's contact details.

4. Call dispatch_communication_tool to record the notification in FHIR. Pass:
   service_request_id, patient_id, the practitioner_id from step 3, the
   acr_category, and a one-sentence finding_summary from the report impression.

5. Call track_acknowledgment_tool with action="create" to start the ack
   countdown. Use 60 minutes for Cat1, 1440 minutes (24 hours) for Cat2.

6. If asked to check on a Task, call track_acknowledgment_tool with
   action="check". If overdue, call escalate_tool — pass the original_task_id
   plus the same study details. This notifies the on-call provider and opens
   a new Task.

7. At any point, call query_audit_tool to return the full Communication and
   Task history for a service_request_id or patient_id.

Always confirm the result of each tool call in your response. If a tool returns
an error, surface it clearly and do not proceed.
"""


def build_agent():
    """Build the ADK Agent. Imported lazily so the module can be loaded
    without google-adk installed (useful for unit tests)."""
    try:
        from google.adk.agents import Agent
    except ImportError as e:
        raise RuntimeError(
            "google-adk is required to build the agent. "
            "Install with: pip install google-adk"
        ) from e

    model = os.getenv("CRITCOM_LLM_MODEL", "gemini-2.0-flash")
    extension_uri = os.getenv(
        "CRITCOM_FHIR_EXTENSION_URI",
        "https://promptopinion.ai/schemas/a2a/v1/fhir-context",
    )

    agent = Agent(
        name="critcom",
        model=model,
        description="Critical results communication agent for radiology",
        instruction=INSTRUCTION,
        tools=ALL_TOOLS,
        before_model_callback=make_extract_fhir_context(extension_uri),
    )
    return agent


# Module-level instance — only built when ADK is available
try:
    root_agent = build_agent()
except Exception:  # noqa: BLE001
    root_agent = None  # type: ignore[assignment]
