requirements.txt
# Core
pydantic>=2.6
pydantic-settings>=2.0
httpx>=0.27
structlog>=24.1
tenacity>=8.0
python-dotenv>=1.0

# LLM (Gemini — used by both the agent and the ACR classifier)
google-generativeai>=0.8

# Agent / A2A
google-adk>=0.1.0
# Pin: google-adk 1.32 imports a2a.server.apps which moved/was removed in a2a-sdk 1.0
a2a-sdk>=0.3,<1.0
starlette>=0.37
uvicorn[standard]>=0.27

# NOTE: pydicom and pynetdicom are intentionally excluded.
# This build uses the FHIR-only workflow for Prompt Opinion deployment.
